// frontend/src/api/generated/models/CustomerDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CustomerDto = {
    id?: number;
    code?: string | null;
    name?: string | null;
    contact?: string | null;
    phone?: string | null;
    createdAt?: string;
    createdBy?: string | null;
    updatedAt?: string | null;
    updatedBy?: string | null;
};